fixed some memory holes, seems to run on linux
git-svn-id: file:///home/mb312/repositories/public-repo/bic/trunk@1333 c5881a9f-4ef2-0310-b211-a1cdbd18b07b

+++ misc_tools/Recon/data.h
@@ -3,38 +3,38 @@
 
 
 void nrerror(char error_text[]);
-float *vector(long np);
-int *ivector(long np);
-double *dvector(long np);
-float **matrix(long nrow, long ncol);
-double **dmatrix(long nrow, long ncol);
+float *vector(int np);
+int *ivector(int np);
+double *dvector(int np);
+float **matrix(int nrow, int ncol);
+double **dmatrix(int nrow, int ncol);
 double **dmatrix_colmajor(int nrow, int ncol);
-fftw_complex **cmatrix(long nrow, long ncol);
-fftw_complex **cmatrix_colmajor(long nrow, long ncol);
-float **submatrix(float **a, long oldrl, long oldrh, long oldcl, long oldch,
-		  long newrl, long newcl);
-void free_vector(float *v, long nl, long nh);
-void free_ivector(int *v, long nl, long nh);
-void free_dvector(double *v, long nl, long nh);
+fftw_complex **cmatrix(int nrow, int ncol);
+fftw_complex **cmatrix_colmajor(int nrow, int ncol);
+float **submatrix(float **a, int oldrl, int oldrh, int oldcl, int oldch,
+		  int newrl, int newcl);
+void free_vector(float *v, int nl, int nh);
+void free_ivector(int *v, int nl, int nh);
+void free_dvector(double *v, int nl, int nh);
 void free_matrix(float **m);
 void free_dmatrix(double **m);
 void free_cmatrix(fftw_complex **m);
-void free_submatrix(float **b, long nrl, long nrh, long ncl, long nch);
+void free_submatrix(float **b, int nrl, int nrh, int ncl, int nch);
 void free_d3tensor(double ***t);
 void free_c3tensor(fftw_complex ***t);
 void free_c4tensor(fftw_complex ****t);
 void free_d4tensor(double ****t);
 void free_zarray(fftw_complex *t);
 
-double ***d3tensor_alloc(long nsl, long nrow, long ncol);
-fftw_complex ***c3tensor_alloc(long nsl, long nrow, long ncol);
-fftw_complex ****c4tensor_alloc(long nvol, long nsl, long nrow, long ncol);
-double ****d4tensor_alloc(long nvol, long nsl, long nrow, long ncol);
+double ***d3tensor_alloc(int nsl, int nrow, int ncol);
+fftw_complex ***c3tensor_alloc(int nsl, int nrow, int ncol);
+fftw_complex ****c4tensor_alloc(int nvol, int nsl, int nrow, int ncol);
+double ****d4tensor_alloc(int nvol, int nsl, int nrow, int ncol);
 
-fftw_complex *zarray(long dsize);
+fftw_complex *zarray(int dsize);
 
 /* Definitions for typical complex->real data conversions */
-double *mag(double *r, const fftw_complex *z, long N);
-double *angle(double *r, const fftw_complex *z, long N);
-double *real(double *r, const fftw_complex *z, long N);
-double *imag(double *r, const fftw_complex *z, long N);
+double *mag(double *r, const fftw_complex *z, int N);
+double *angle(double *r, const fftw_complex *z, int N);
+double *real(double *r, const fftw_complex *z, int N);
+double *imag(double *r, const fftw_complex *z, int N);
